first implementation of block refactor
