+++ datatypes/src/main/java/org/hyperledger/besu/datatypes/Transaction.java
@@ -213,6 +213,9 @@ public interface Transaction {
    */
   Optional<List<AccessListEntry>> getAccessList();
 
+  int getBlobCount();
+
+
   /**
    * Returns the transaction with the proper encoding
    *
